GRAILS-10091 - Improve service bean aliasing
The framework is only supposed to create aliased bean names for services if the alias name will not conflict with another bean.  This works if the conflicting bean provided by the application is a Service artifact.  It does not work if the conflict is created by a bean that is not a Service artifact but is a bean explicitly added to the context.  This test represents that scenario and is currently marked with @NotYetImplemented.